% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fechas.R
\name{add_delimiter}
\alias{add_delimiter}
\title{Add delimiter to date}
\usage{
add_delimiter(dates)
}
\arguments{
\item{dates}{A character vector of dates with no delimiter.}
}
\value{
A character vector with the dates having the '/' delimiter.
}
\description{
Adds the '/' delimiter to a vector of dates with no delimiter. This will make it easier to standardize these types of dates using the standardize_dates function.
}
\examples{
add_delimiter(c("01012020", "22032023", "220323", "20200101"))
}
