% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitivas.R
\name{Consistency_Format_Cleanup}
\alias{Consistency_Format_Cleanup}
\alias{primitives}
\alias{Syntactic_Accuracy_Cleanup}
\alias{Accuracy_Range_Cleanup}
\alias{Record_Completeness_Cleanup}
\alias{Referential_Integrity_Cleanup}
\alias{Risk_Inconsistency_Cleanup}
\title{Primitives to improve data quality}
\usage{
Consistency_Format_Cleanup(ds, cleaningMethod, ...)

Syntactic_Accuracy_Cleanup(ds, cleaningMethod, ...)

Accuracy_Range_Cleanup(ds, cleaningMethod, ...)

Record_Completeness_Cleanup(ds, cleaningMethod, ...)

Referential_Integrity_Cleanup(ds, cleaningMethod, ...)

Risk_Inconsistency_Cleanup(ds, cleaningMethod, ...)
}
\arguments{
\item{ds}{The data set whose quality is to be improved.}

\item{cleaningMethod}{The cleaning method to be applied.}

\item{...}{Additional arguments to be passed to the chosen cleaning method.}
}
\value{
The return value of the cleaning method used or an error message if
  an invalid cleaning method is chosen.
}
\description{
These primitives add a higher abstraction layer for the use of data
cleaning methods, and relate them directly to the data quality properties
that they affect.
}
\examples{
ds <- data.frame(dates = c("2021-01-01", "02-01-2021", "03/27/2023"))
Consistency_Format_Cleanup(ds, "standardize_dates")
}
\seealso{
For more information about the cleaning methods available for each
  property (and therefore for each primitive), see the
  \code{\link[=properties]{following list}}.
}
