% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integridad.R
\name{derived_column}
\alias{derived_column}
\title{Creates a new column in a data frame using a formula}
\usage{
derived_column(df, new_col, formula)
}
\arguments{
\item{df}{The data frame where the new column will be added.}

\item{new_col}{A character string with the name of the new column to be
created.}

\item{formula}{A formula that specifies how to calculate the new column based
on other columns in the data frame. The formula can include arithmetic
operations and function calls, and can reference other columns in this data
frame (df) or another data frame using their names or indices.}
}
\value{
The same data frame with the new column added.
}
\description{
This function creates a new column in a data frame by evaluating a formula
that references other columns in this (df) or other data frame. The new
column is added to the data frame in-place.
}
\examples{
data <- data.frame(x = 1:5, y = 6:10)
derived_column(data, "z", quote(x + y))
}
