% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{find_na}
\alias{find_na}
\title{Find NA}
\usage{
find_na(vector)
}
\arguments{
\item{vector}{A vector.}
}
\value{
A vector of the positions of NA values in vector.
}
\description{
Finds the positions of NA values in a vector.
}
\examples{
find_na(c(1, NA, 2, NA))
}
