% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitivas.R
\name{Limpieza_Consistencia_Formato}
\alias{Limpieza_Consistencia_Formato}
\alias{primitives}
\alias{Limpieza_Exactitud_Sintactica}
\alias{Limpieza_Rango_Exactitud}
\alias{Limpieza_Completitud_Registro}
\alias{Limpieza_Integridad_Referencial}
\alias{Limpieza_Riesgo_Inconsistencia}
\title{Primitives to improve data quality}
\usage{
Limpieza_Consistencia_Formato(ds, funcionLimpieza, ...)

Limpieza_Exactitud_Sintactica(ds, funcionLimpieza, ...)

Limpieza_Rango_Exactitud(ds, funcionLimpieza, ...)

Limpieza_Completitud_Registro(ds, funcionLimpieza, ...)

Limpieza_Integridad_Referencial(ds, funcionLimpieza, ...)

Limpieza_Riesgo_Inconsistencia(ds, funcionLimpieza, ...)
}
\arguments{
\item{ds}{The data set whose quality is to be improved.}

\item{funcionLimpieza}{The cleaning method to be applied.}

\item{...}{Additional arguments to be passed to the chosen cleaning method.}
}
\value{
The return value of the cleaning method used or an error message if
  an invalid cleaning method is chosen.
}
\description{
These primitives add a higher abstraction layer for the use of data
cleaning methods, and relate them directly to the data quality properties
that they affect.
}
\note{
For non-Spanish speakers, the name of the primitives contains the name
  of the data quality property they deal with in Spanish. To find out which
  one to use, you can refer to the link included in the "see also" section.
  There you can find the properties in both English and Spanish.
}
\examples{
ds <- data.frame(dates = c("2021-01-01", "02-01-2021", "03/27/2023"))
Limpieza_Consistencia_Formato(ds, "standardize_dates")
}
\seealso{
For more information about the cleaning methods available for each
  property (and therefore for each primitive), see the
  \code{\link[=properties]{following list}}.
}
