% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completitud.R
\name{na_detection}
\alias{na_detection}
\title{Detect rows with missing values in a data set}
\usage{
na_detection(data, columns)
}
\arguments{
\item{data}{The data set on which to detect missing values.}

\item{columns}{A vector with the indices of the columns to be checked.}
}
\value{
A vector with the indices of the rows that have missing values in the
  specified columns.
}
\description{
This function takes a data set and a vector with the indices of the columns
to be checked for missing values. It returns a vector with the indices of the
rows that have at least one missing value in any of the specified columns.
}
\note{
This function aims to help improve the property
  \code{\link[=properties]{Record Completeness}}.
}
\examples{
# Example data frame
users <- data.frame(idUser = c(1, 2, NA, 4, 5),
                    name = c("Mario", "Ismael", "Fernando", "Gregorio", "John"),
                    tfno = c("602364787", NA, "603157489", "650879511", NA))
users

# We obtain the indexes of the rows containing any nulls in the idUser or tfno columns
na_rows <- na_detection(users, c(1, 3))

# We create a data frame with only the valid records
valid_rows <- users[-na_rows,]
valid_rows

# We create a data frame with only the invalid records
invalid_rows <- users[na_rows,]
invalid_rows
}
\seealso{
To impute the NA values, check the function \code{\link{na_imputation}}.
}
