% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completitud.R
\name{na_imputation}
\alias{na_imputation}
\title{Impute missing values}
\usage{
na_imputation(
  df,
  column,
  method,
  target = NULL,
  seed = NULL,
  print_flag = TRUE,
  no_attrs = FALSE
)
}
\arguments{
\item{df}{A data frame.}

\item{column}{Name or index of the column to impute missing values.}

\item{method}{String specifying the imputation method. Possible
values are "mean", "median", "mode", "knn", "rf", and "mice".}

\item{target}{Target variable. Defaults to NULL.}

\item{seed}{An integer specifying the seed to use in mice (used only in
"mice" method). Defaults to NULL.}

\item{print_flag}{Logical indicating if mice will print running log on
console (used only when method is "mice"). Use \code{print_flag = FALSE}
for silent computation. Defaults to TRUE.}

\item{no_attrs}{Logical indicating whether to return numerical variable or
categorical variable (TRUE), or the imputation class (FALSE). Defaults to
FALSE.}
}
\value{
If \code{no_attrs = FALSE}, then return imputation class. If
  \code{no_attrs = TRUE}, then return numerical vector or factor instead.
}
\description{
This function imputes missing values in a data frame column using various
imputation methods. It uses the \code{\link{imputate_na}()} function from the
\code{\link{dlookr}} package to impute the missing values.
}
\note{
This function aims to help improve the property
  \code{\link[=characteristics_properties]{Record Completeness}}.
}
\examples{
# Generate data for the example
heartfailure2 <- dlookr::heartfailure
heartfailure2[sample(seq(NROW(heartfailure2)), 20), "platelets"] <- NA

# Replace the missing value of the platelets variable with median
na_imputation(heartfailure2, "platelets", method = "median")
}
\seealso{
For more information about the function, check the \code{\link{imputate_na}}
documentation.

To get a visualization and/or an overview of the returned value, check the
\code{\link{plot_imputation}} and \code{\link{summary_imputation}} functions
respectively.

To simply detect NA values, check the function \code{\link{na_detection}}.
}
