% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangos.R
\name{outside_range}
\alias{outside_range}
\title{Detect values outside range}
\usage{
outside_range(vector, min, max, value = FALSE)
}
\arguments{
\item{vector}{A numeric vector to be evaluated.}

\item{min}{The lower bound of the range.}

\item{max}{The upper bound of the range.}

\item{value}{Logical indicating whether to return the values outside of the
range instead of their indices (default: FALSE).}
}
\value{
A numeric vector of either the values or indices that are outside of
  the specified range.
}
\description{
Given a vector of numeric values, returns the indices or values that are
outside of the specified range.
}
\note{
This function aims to help improve the property
  \code{\link[=properties]{Accuracy Range}}.
}
\examples{
# Example data frame
employees <- data.frame(idUser = c(1, 2, 3, 4, 5),
                        name = c("Mario", "Ismael", "Fernando", "Gregorio", "John"),
                        salaries = c(10000, 3600, 4000, 500, 800))
employees

# Minimum wage
min <- 900

# Maximum wage
max <- 5000

### Using value = FALSE
# We obtain the indexes of the invalid salaries from the salaries column
invalid_indexes <- outside_range(employees$salaries, min, max)

# We load the wrong values
invalid_salaries <- employees$salaries[invalid_indexes]
invalid_salaries

### Using value = TRUE
# We obtain the invalid salaries
invalid_salaries <- outside_range(employees$salaries, min, max, value = TRUE)
invalid_salaries

# We create a data frame with only the valid records
valid_employees <- employees[!employees$salaries \%in\% invalid_salaries,]
valid_employees # or employees[employees$salaries >= min & employees$salaries <= max,]

# We create a data frame with only the invalid records
invalid_employees <- employees[employees$salaries \%in\% invalid_salaries,]
invalid_employees # or employees[employees$salaries < min | employees$salaries > max,]
}
\seealso{
\code{\link{check_categories}}
}
