% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadenas.R
\docType{data}
\name{patterns}
\alias{patterns}
\title{Examples of Common Patterns}
\format{
A data frame with 4 rows and 3 columns:
\describe{
  \item{Name}{Name of the pattern}
  \item{Pattern}{Regular expression used to validate the element}
  \item{Example}{An example of a valid value of the element}
}
}
\usage{
data(patterns)

patterns
}
\description{
Examples of regular expressions of common elements to use with the functions
\code{\link{check_regex}} or \code{\link{check_regexl}}
}
\examples{
dni <- c("71684024L", "5684544A", "05331975M", "52498742", "603489758")
data(patterns)
dni_pattern <- patterns[patterns$Name == "DNI", 2]
check_regexl(dni_pattern, dni)
}
\keyword{datasets}
