% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{plot_imputation}
\alias{plot_imputation}
\title{Visualize Information for an "imputation" Object}
\usage{
plot_imputation(x, typographic = TRUE, base_family = NULL)
}
\arguments{
\item{x}{An object of class "imputation", result of a call to
na_imputation().}

\item{typographic}{Logical indicating whether to apply focuses on typographic
elements to ggplot2 visualization. The default is TRUE. If TRUE provides a
base theme that focuses on typographic elements using hrbrthemes package.}

\item{base_family}{String indicating the name of the base font family to use
for the visualization. If not specified, the font defined in dlookr is
applied.}
}
\value{
Graph to visualize \code{x}.
}
\description{
This function was made to plot the "imputation" class objects result of a
call to \code{\link{na_imputation}}. It can visualize two kinds of plot: the
imputation of a numerical variable is a density plot, and the imputation of a
categorical variable is a bar plot.
}
\examples{
# Generate data for the example
heartfailure2 <- dlookr::heartfailure
heartfailure2[sample(seq(NROW(heartfailure2)), 20), "platelets"] <- NA

# Replace the missing value of the platelets variable with median
imputation <- na_imputation(heartfailure2, "platelets", method = "median")

plot_imputation(imputation, typographic = FALSE)
}
\seealso{
For more information about the function, check the
\code{\link{plot.imputation}} documentation.

Related functions: \code{\link{na_imputation}, \link{summary_imputation}}
}
