% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{replace_values}
\alias{replace_values}
\title{Replace values}
\usage{
replace_values(vector, positions, replacement)
}
\arguments{
\item{vector}{A vector.}

\item{positions}{The positions to replace.}

\item{replacement}{The values to replace with.}
}
\value{
Returns a vector with the specified values replaced.
}
\description{
Replace values in a vector.
}
\examples{
replace_values(1:5, c(2,4), c(0, 7))
}
