% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fechas.R
\name{standardize_dates}
\alias{standardize_dates}
\title{Standardize dates}
\usage{
standardize_dates(
  dates,
  input_format = c("ymd", "mdy", "dmy"),
  output_format = "\%Y-\%m-\%d"
)
}
\arguments{
\item{dates}{A vector of dates to be standardized}

\item{input_format}{The input format of the dates. If not specified, the function will try to parse the dates in the following formats: ymd, mdy, dmy.}

\item{output_format}{The output format of the dates. The default is "yyyy-mm-dd".}
}
\value{
A vector of standardized dates.
}
\description{
Converts a vector of dates from one or more formats to another.
}
\examples{
standardize_dates(c("12/31/2019", "27-03-20"))
}
