% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fechas.R
\name{standardize_dates_table}
\alias{standardize_dates_table}
\title{Standardize Dates Table}
\usage{
standardize_dates_table(
  dates,
  columns,
  input_format = c("ymd", "mdy", "dmy"),
  output_format = "\%Y-\%m-\%d"
)
}
\arguments{
\item{dates}{A dataframe containing dates to be standardized.}

\item{columns}{The columns to standardize in the dataframe.}

\item{input_format}{The format of the dates in the dataframe. Defaults to: ymd, mdy, dmy.}

\item{output_format}{The output format of the dates. Defaults to "yyyy-mm-dd".}
}
\value{
The same dataframe with the specified columns standardized.
}
\description{
Takes a dataframe and converts the values in the specified columns from one format to another.
}
\examples{
# Create dataframe with dates
dates <- data.frame(id = 1:2,
                    date1 = c("Feb 2, 2021", "02032021"),
                    date2 = c("12/31/2019", "27-03-20"))

# Standardize dates
dates <- standardize_dates_table(dates, c(2, 3))

}
