% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{summary_imputation}
\alias{summary_imputation}
\title{Summarizing imputation information}
\usage{
summary_imputation(object)
}
\arguments{
\item{object}{An object of class "imputation", result of a call to
na_imputation().}
}
\value{
Information about \code{object}.
}
\description{
This function was made to print and summarize the "imputation" class objects
result of a call to \code{\link{na_imputation}}.
}
\examples{
# Generate data for the example
heartfailure2 <- dlookr::heartfailure
heartfailure2[sample(seq(NROW(heartfailure2)), 20), "platelets"] <- NA

# Replace the missing value of the platelets variable with median
imputation <- na_imputation(heartfailure2, "platelets", method = "median")

summary_imputation(imputation)
}
\seealso{
For more information about the function, check the
\code{\link{summary.imputation}} documentation.

Related functions: \code{\link{na_imputation}, \link{plot_imputation}}
}
