% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{find_na}
\alias{find_na}
\title{Find NA values}
\usage{
find_na(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector with the positions of the NA elements of x.
}
\description{
This function was made for the purpose of locating the possible NA values of
the vector returned by \code{\link{standardize_dates}()}. These would be
values that could not be parsed. Thanks to this you can access the original
values of the date vector and fix them manually. Once fixed,
\code{\link{replace_values}()} can be used to replace these values in the
resulting vector.
}
\examples{
find_na(c(1, NA, 2, NA))
}
\seealso{
\code{\link{standardize_dates}, \link{replace_values}}
}
