% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{replace_values}
\alias{replace_values}
\title{Replace values in a vector}
\usage{
replace_values(x, positions, replacement)
}
\arguments{
\item{x}{A vector.}

\item{positions}{The positions to replace.}

\item{replacement}{The values to replace with.}
}
\value{
Returns the introduced vector with the specified values replaced.
}
\description{
This function was made to replace the NA values returned by
standardize_dates() with the manually fixed values.
}
\examples{
replace_values(1:5, c(2,4), c(0, 7))

}
