% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integridad.R
\name{verify_integrity}
\alias{verify_integrity}
\title{Verifies referential integrity between two data frames}
\usage{
verify_integrity(df1, df2, col1, col2)
}
\arguments{
\item{df1}{The first data frame to compare.}

\item{df2}{The second data frame to compare.}

\item{col1}{The name of the column in the first data frame (df1) that
contains the primary values.}

\item{col2}{The name of the column in the second data frame (df2) that
contains the foreign values.}
}
\value{
A vector with the values that violate referential integrity. If no
  violated values are found, an empty vector is returned.
}
\description{
This function verifies whether the values in one column of a data frame (df2)
are present in another column of a different data frame (df1), to ensure
referential integrity between both data frames.
}
\note{
This function aims to help improve the property
  \code{\link[=characteristics_properties]{Referential Integrity}}.
}
\examples{
employees <- data.frame(id = 1:5, name = c("Mario", "Ismael", "Fernando", "Gregorio", "John"))
sales <- data.frame(id = 1:5, price = c(500, 1000, 250, 300, 500),
                    date = c("08/01/2022", "16/03/2022", "03/04/2022", "04/04/2022", "10/05/2022"),
                    seller = c(1, 2, 3, 3, 6))

verify_integrity(employees, sales, "id", "seller")
}
